requirements.txt
asgiref==3.8.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-polymorphic==3.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-polymorphic==2.0.0
drf-spectacular==0.28.0
httpie==3.2.4
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
pillow==11.2.1
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
